+++ build_runner/lib/src/server/server.dart
@@ -46,11 +46,11 @@ ServeHandler createServeHandler(WatchImpl watch) {
   var rootPackage = watch.packageGraph.root.name;
   var assetGraphHanderCompleter = Completer<AssetGraphHandler>();
   var assetHandlerCompleter = Completer<AssetHandler>();
-  watch.ready.then((_) async {
-    assetHandlerCompleter.complete(AssetHandler(watch.reader, rootPackage));
-    assetGraphHanderCompleter.complete(
-        AssetGraphHandler(watch.reader, rootPackage, watch.assetGraph!));
-  });
+  watch.reader.then((reader) async {
+    assetHandlerCompleter.complete(AssetHandler(reader, rootPackage));
+    assetGraphHanderCompleter
+        .complete(AssetGraphHandler(reader, rootPackage, watch.assetGraph!));
+  }).catchError((_) {}); // These errors are separately handled.
   return ServeHandler._(watch, assetHandlerCompleter.future,
       assetGraphHanderCompleter.future, rootPackage);
 }
@@ -164,6 +164,7 @@ class ServeHandler implements BuildState {
 
   Future<shelf.Response> _assetsDigestHandler(
       shelf.Request request, String rootDir) async {
+    final reader = await _state.reader;
     var assertPathList =
         (jsonDecode(await request.readAsString()) as List).cast<String>();
     var rootPackage = _state.packageGraph.root.name;
@@ -171,7 +172,7 @@ class ServeHandler implements BuildState {
     for (final path in assertPathList) {
       try {
         var assetId = pathToAssetId(rootPackage, rootDir, p.url.split(path));
-        var digest = await _state.reader.digest(assetId);
+        var digest = await reader.digest(assetId);
         results[path] = digest.toString();
       } on AssetNotFoundException {
         results.remove(path);
@@ -216,9 +217,10 @@ class BuildUpdatesWebSocketHandler {
 
   Future emitUpdateMessage(BuildResult buildResult) async {
     if (buildResult.status != BuildStatus.success) return;
-    var digests = <AssetId, String>{};
+    final reader = await _state.reader;
+    final digests = <AssetId, String>{};
     for (var assetId in buildResult.outputs) {
-      var digest = await _state.reader.digest(assetId);
+      var digest = await reader.digest(assetId);
       digests[assetId] = digest.toString();
     }
     for (var rootDir in connectionsByRootDir.keys) {